Iconv tool (#160)
* First implementation with only simdutf functions.

* CMake files.

* Better run_procedure.

* Iconv fallback.

* Filesystem header issue.

* Warnings for validate_with_errors scalar and string issues.

* Fix error with iconv message and alias UTF-16 for UTF-16LE.

* Patch.

* Patch.

* Garbage

* Getting good results.

* Removing dead comments.

* Removing bogus commenting

* Adding string header

* Cleaner check.

* CMake linking.

* Attempt to fix string issues.

* Link iconv and rewrite iconv_fallback.

* Add guards when iconv is not available.

* Typo.

* Executable instead of library.

* Alias UTF-32 for UTF-32LE

* Fix iconv_fallback.

* Warning fix.

* Patch to reduce exec size.

* First work.

* One-chunk loading working.

* Error messages and typos.

* Simdutf streaming.

* Iconv fallback streaming.

* Use one output buffer.

* Change to std::array for output_buffer

* Fix from utf8.

* Fix utf-16 truncation across chunks.

* Format specifier printf.

* Change CHUNK_SIZE to 64kb.

* %zu specifier.

* Remove leftovers for final iterations.

* More options and clearer user interface.

* README.

+++ src/scalar/utf32.h
@@ -20,7 +20,7 @@ inline simdutf_warn_unused bool validate(const char32_t *buf, size_t len) noexce
 
 inline simdutf_warn_unused result validate_with_errors(const char32_t *buf, size_t len) noexcept {
   const uint32_t *data = reinterpret_cast<const uint32_t *>(buf);
-  uint64_t pos = 0;
+  size_t pos = 0;
   for(;pos < len; pos++) {
     uint32_t word = data[pos];
     if(word > 0x10FFFF) {
